fix(Router): change RequestContext into a RoutingConditionService
