+++ src/Resources/config/services.yaml
@@ -40,11 +40,6 @@ services:
             $dataClass: '%monsieurbiz_cms_page.model.page_translation.class%'
             $validationGroups: ['monsieurbiz']
 
-    # Routing Context
-    MonsieurBiz\SyliusCmsPagePlugin\Routing\RequestContext:
-        decorates: router.request_context
-        arguments: ['@MonsieurBiz\SyliusCmsPagePlugin\Routing\RequestContext.inner']
-
     monsieurbiz.cms_page.datetime_provider:
         class: Symfony\Component\Clock\Clock
         public: true
