+++ src/Routing/PageSlugConditionChecker.php
@@ -17,8 +17,11 @@ use MonsieurBiz\SyliusCmsPagePlugin\Repository\PageRepositoryInterface;
 use Sylius\Component\Channel\Context\ChannelContextInterface;
 use Sylius\Component\Channel\Context\ChannelNotFoundException;
 use Sylius\Component\Locale\Context\LocaleContextInterface;
+use Symfony\Bundle\FrameworkBundle\Routing\Attribute\AsRoutingConditionService;
 use Symfony\Component\Clock\ClockInterface;
+use Symfony\Component\HttpFoundation\Request;
 
+#[AsRoutingConditionService(alias: 'monsieurbiz.cms_page.route_checker')]
 final class PageSlugConditionChecker
 {
     public function __construct(
@@ -27,6 +30,11 @@ final class PageSlugConditionChecker
         private LocaleContextInterface $localeContext,
         private ClockInterface $clock,
     ) {
+    }
+
+    public function checkPageSlug(Request $request): bool
+    {
+        return $this->isPageSlug($this->prepareSlug($request->getPathInfo()));
     }
 
     public function isPageSlug(string $slug): bool
@@ -41,5 +49,18 @@ final class PageSlugConditionChecker
         } catch (ChannelNotFoundException $channelNotFoundException) {
             return false;
         }
+    }
+
+    private function prepareSlug(string $slug): string
+    {
+        $slug = ltrim($slug, '/');
+        $localeCode = $this->localeContext->getLocaleCode();
+
+        if (false === strpos($slug, $localeCode)) {
+            return $slug;
+        }
+
+        // Remove the locale code which is at the beginning of the slug
+        return (string) preg_replace(\sprintf('/^%s\//', $localeCode), '', $slug);
     }
 }
